Login updated according to new template.
