+++ implementation/frontend/material-dashboard-react-v1.5.0-retro/src/assets/css/material-dashboard-react.css
@@ -74,6 +74,24 @@ small {
   font-size: 80%;
 }
 
+white {
+  color: "rgba(255,255,255,.62)";
+  margin: "0";
+  fontSize: "14px";
+  marginTop: "0";
+  marginBottom: "0";
+}
+
+titleWhite {
+  color: "#FFFFFF";
+  marginTop: "0px";
+  minHeight: "auto";
+  fontWeight: "300";
+  fontFamily: "'Roboto', 'Helvetica', 'Arial', sans-serif";
+  marginBottom: "3px";
+  textDecoration: "none";
+}
+
 h1 {
     font-size: 3em;
     line-height: 1.15em;
