+++ implementation/frontend/material-dashboard-react-v1.5.0-retro/src/routes/index.jsx
@@ -1,5 +1,5 @@
 import Dashboard from "layouts/Dashboard/Dashboard.jsx";
-import Login from "views/Login/Login.js";
+import Login from "views/Login/Login.jsx";
 
 // TODO: add if statement so login page shown.
 const indexRoutes = [{ path: "/", component: Dashboard }];
