--- implementation/frontend/material-dashboard-react-v1.5.0-retro/src/views/Login/Login.css
@@ -1,52 +1,0 @@
-@media all and (min-width: 480px) {
-  .Login {
-  }
-
-  .Retro-logo {
-    position: fixed;
-    left: 50%;
-    right: 50%;
-    z-index: 2;
-    top: 0;
-    text-align: center;
-    font-size: 800%;
-  }
-
-  .Login-background {
-    position: fixed;
-    left: 0;
-    right: 0;
-    z-index: 1;
-    background: #FFFAFF;
-    background-size: cover;
-  }
-
-  .Login-title {
-    color: #C93745;
-    text-align: center;
-    font-family: "Comic Sans MS", cursive, sans-serif;
-    font-size: 40px;
-    align-items: center;
-    background: #ffffff;
-  }
-
-  .Login-box {
-    border-style: double;
-    background: #FFF9FF;
-    position: fixed;
-    left: 0;
-    right: 0;
-    z-index: 1;
-    width: 20%;
-    position: relative;
-    top: 50px;
-    left: 200px;
-    align-items: center;
-  }
-
-  .Login-text {
-    font-family: "Comic Sans MS", cursive, sans-serif;
-    font-size: 15px;
-    align-items: center;
-  }
-}
